requirements.txt
absl-py==0.15.0
astunparse==1.6.3
attrs==21.2.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
clang==5.0
click==8.0.3
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
h5py==3.1.0
idna==3.3
imageio==2.13.3
importlib-metadata==4.8.2
joblib==1.1.0
jsonschema==3.2.0
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.6
matplotlib==3.3.4
networkx==2.5.1
notifiers==1.2.1
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.4.60
opt-einsum==3.3.0
pandas==1.1.5
Pillow==8.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyWavelets==1.1.1
requests==2.26.0
requests-oauthlib==1.3.0
rfc3987==1.3.8
rsa==4.8
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
six==1.15.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.2
tensorflow-estimator==2.6.0
termcolor==1.1.0
threadpoolctl==3.0.0
tifffile==2020.9.3
typing-extensions==3.7.4.3
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.12.1
zipp==3.6.0
